Add sample Vim ftplugin integration
This file demonstrates a lot of options for integration. I think that's
useful but it means the file itself is unsuitable for immediate usage
even though it works right now. Challenges with merge commits and gq
make the naive integration uncomfortable but at the same time it's
difficult to make a "recommended" setup.

Alternatively, there could be several sample integrations suitable for
different needs, with either no default or a "sensible" default. At that
point, though, a stand-alone Vim plugin with options to control
behaviour might be far better; it would certainly be easier to install.