Add basic GitLab CI integration
Add a basic GitLab CI configuration based on the official sample
configuration [1]. Note that "cargo test" does not build a binary, so
the tests that rely on a binary fail without an explicit build action.

There are more advanced alternatives, such as more support for different
toolchain versions [2] and potential for execution speed reduction [3].
The former is nice in principle but uses its own Docker images; and
that's probably fine but not as desirable as an official image. The
latter may not be relevant yet; the build presently runs in about 2
minutes and I've previously found GitLab's CI cache unreliable.

Most interesting is "trust" [4], whose claim to fame is multi-platform
compilation with binary distribution, but it doesn't support GitLab.

[1] https://gitlab.com/gitlab-org/gitlab-ce/blob/1553a34dbff167978f5dc81cc3a21e0b3b2b2bfa/vendor/gitlab-ci-yml/Rust.gitlab-ci.yml
[2] https://imp.gitlab.io/blog/blog/gitlab-ci-and-rust/
[3] https://blog.wjlr.org.uk/2016/08/16/fast-rust-gitlab-ci.html
[4] https://github.com/japaric/trust