Provide precompiled binaries via japaric/trust
Integrate Travis CI and AppVeyor to provide precompiled binaries, hosted
on GitHub, using the "trust" template [1] as a basis. Keep the current
GitLab CI integration as the primary CI, running the others only
post-integration and uploading artifacts only for tags. GitLab's API
provides no means with which to attach assets to "releases" so this is
only possible to do on GitHub; to that end I need to mirror the
repository. GitLab is still the better host so migrating is of no
interest.

Build for 64-bit Linux, Windows, macOS, and BSD until somebody needs
more than this.

The distribution contains the binary, the license, the readme, and the
contrib/ directory, all nested inside a directory for unpacking
convenience. The binary is stripped where necessary.

Main changes to the template:

- All shell scripts are plain Bourne Shell and pass ShellCheck 0.4.6;
  the biggest offender was strictly-unnecessary quoting.
- Primary CI is elsewhere so we don't need to spend energy on debug
  builds.
- The distribution contains more stuff.

[1] https://github.com/japaric/trust