Add man page
Compile a man page from AsciiDoc with Asciidoctor and package it in the
distribution on all platforms.

I've experimented with compilation inside a build.rs program but had
trouble getting this working reliably across three different platforms
and container configurations, at least in part, but I suspect not
solely, because of non-existent absolute paths. After struggling with
that for too long I realised there was absolutely no reason to tie the
two processes together.