Fix man page build
In the process of making commit 28743e8 ("Add man page", 2018-05-24) the
"doc" directory momentarily existed as "docs". I located all references
to the latter using rg(1) and updated them when the directory was
renamed. Unfortunately I didn't know what I was doing: rg skips hidden
files by default and the Travis CI specification happens to be hidden.
In Travis, Asciidoctor still looks in the original path, which is
invalid, and causes the build to fail.