Support explicit core.commentChar
ref #3

If the core.commentChar is set to a specific character, use that to
denote a comment line. If it is unset, use the default value of "#". If
it is set to "auto", also use the default value; "auto" support can be
added on-demand.

The "auto" algorithm was implemented in git@84c9dc2c5a and is a lot
simpler than the documentation suggests: it looks for 10 candidate ASCII
characters in the current message's first column, disqualifying any that
are found and erroring out if no candidates remain viable. This is
straight-forward to replicate but requires an extra pass so I don't want
to do it unless I have to.