Update sample Vim configuration
I often paste a URL as a reference. With non-zero 'textwidth', a URL
exceeding the configuration value is pushed to the next line.
commitmsgfmt cannot undo this. Since commitmsgfmt is supposed to be the
authority on wrapping there is no reason to also use Vim for wrapping,
especially not when Vim undermines commitmsgfmt.