Only wrap on space
When wrapping a line, only do so at the last space before the width
limit. If the line contains no spaces, effectively forming a single long
word, do not wrap the line at all; allow it to extend beyond the limit.
When this situation occurs there is a very high probability it was
deliberate, such as in the case of URLs, and that breaking the "word" is
in fact detrimental. On the off-chance that it was not deliberate and
the line was just a length of unbroken characters, well, there is no
meaningful way to wrap such a line in the first place.

This change alone yields a 6-8 times speed-up as reported by time(1).
Using the performance analysis [1] as a regression shifts the diffstat
means and proportions down by a few trifle decimals.

[1] https://mkjeldsen.gitlab.io/blog/introducing-commitmsgfmt/