Preserve autosquash subject lines
If a subject line matches a pattern recognized by "git-rebase
--autosquash", preserve it exactly. If it is rewritten in any way, it
may fail to match the original and consequently break "--autosquash".

The Git hook called out this requirement and addressed it by
short-circuiting because commitmsgfmt didn't handle it natively, on
account of the user being able to skip the invocation entirely, like the
hook did. However, if a user chooses to amend such a commit with
explanation of the relative improvement for review purposes, for
instance, commitmsgfmt will most likely run and then may break the
subject line.