Support non-breaking spaces for left-binding tokens
If a space-delimited token ("word") logically belongs with its preceding
word, consider these two words to be a single word and disallow wrapping
between them.

If a word begins with the current comment character, disallow wrapping
immediately before it to avoid the new line degenerating into a comment.

There is no support for tokens that should bind to the right, such as
indefinite articles. I was able to prove that this can be accomplished
but only with some very rough heuristics that I cannot yet justify.

Issue: https://gitlab.com/mkjeldsen/commitmsgfmt/-/issues/5