Fix wrapping of footnote references followed by punctuation
When a footnote reference is immediately followed by punctuation, that
reference will not be merged with the preceding word when wrapping, as
it would have been without punctuation. That is,

    foo [1]
    bar [2].

can mistakenly degenerate into

    foo [1]
    bar
    [2].

This commit resolves that by accepting any unbroken sequence of
arbitrary punctuation on the right as part of that same word. The issue
"persists" where at least one space separates the reference and the
start of the punctuation sequence but we take the spacing as a sign of
deliberate intent and allow that.

Fixes: 37da199 (Support non-breaking spaces for left-binding tokens, 2021-04-04)