Make punctuation non-breaking
If a word is comprised solely of ASCII punctuation, consider it to
belong with the previous word and disallow breaking between those words.
Commit 7c03542 (Fix wrapping of footnote references followed by
punctuation, 2022-02-09) identified this edge case and allowed it under
the assumption that the presence of a space indicates intent. However,
this does not seem like a useful construct, and disallowing it neatly
resolves numerous unlikely but troublesome pathological cases such as

    foo ...
    foo =
    foo =>
    foo =:
    foo ---

This leaves a new edge case where

    foo = 42

can turn into

    foo =
    42

That case is harder to solve because of the potential for ambiguity with
list items so we leave it for now.