Require spacing between paragraphs and footnotes
In case a footnote reference within a paragraph ends up at the start of
a new line, that footnote reference degenerates into a footnote and can
no longer wrap with the remainder of the paragraph. That is,

    foo
    [2] bar

becomes fixed in place. commitmsgfmt since v1.4.0 should not have
produced such pathological output in the first place but missed a few
edge cases, and anyway, the input could come from elsewhere. To counter
this degenerative case, require that blocks of footnotes be preceded by
an empty line; this makes the example above be interpreted as a single
paragraph such that "[2]" instead becomes a footnote reference.

On the other hand, in some cases footnotes will now degenerate into
footnote references.