Remove implementation support for tab-indented list items
The documentation specifies that a tab-indented line is a literal, not a
list item, and that list items may be indented with spaces, yet the list
item implementation allows both tabs and spaces in indentation. The
documentation is quite clear and test backing appears incidental so
bring the implementation in line with the documentation by dropping
support for tab-indented list items.