Remove "clap" dependency
Replace the "clap" frontend with manual parsing, clearing up the last
remaining "easy" redundant allocations. With this, allocations drop a
further 32% for an empty message, from 22.34 KiB to 15.1 KiB, and 17%
for a 3rd quartile message, from 42.84 KiB to 35.66 KiB. Compared to
v1.5.0 the total reduction for both is > 99%, and 25% for Project
Gutenberg's Don Quixote. Runtime is immaterially affected both ways, and
the stripped binary size drops by 47%.

Besides the runtime improvements the "clap" removal also improves the
maintenance health considerably:

- The number of direct and transitive dependencies drops by 22%.
- The debug and release build times drop by 64% respectively 70%.
- The v2.x line has been unmaintained for nearly exactly 2 years and
  superseded by 2 major releases, with more majors "expected".
- "clap" >= v3.x assumes the user hostile "derive" (née "structopt")
  integration should be not only core functionality but the primary
  usage mechanism.
- Notwithstanding the lifecycle issues above, "clap"'s aggressive N-1
  MSRV policy makes it difficult to consume the library effectively.

"clap" is a very powerful, overwhelmingly reliable tool that has served
commitmsgfmt admirably for nearly 6 years and is worthy of consideration
in that light. And now it has outstayed its welcome. If improbable
future extensions to the now manually written parser makes maintenance
unwieldy we can look into "bpaf" [1].

[1] https://github.com/pacak/bpaf