ci: archive release source
Upon release, eagerly provide a stable archive of the release source.

It used to be that GitHub's (and GitLab's?) virtual source code assets
were generated on demand to limit resource consumption. This meant the
assets were not reliably available, and that two instances of the same
asset downloaded long apart could be generated with different
implementations, causing different checksums for the same contents.
Eventually somebody raised an overly dramatic stink about this and
GitHub promised to make the archive generation stable but I don't want
to rely on that promise.